requirements.txt
## Web Application
bokeh==0.12.15
Django==2.0.5
django-rest-framework==0.1.0
djangorestframework==3.8.2
dj-database-url==0.5.0
gunicorn==19.7.1
psycopg2==2.7.4
python-dotenv==0.8.2
social-auth-app-django==2.1.0
social-auth-core==1.7.0
whitenoise==3.3.1

## Utilities
pytz==2018.4
python-dateutil==2.7.2
requests==2.18.4
six==1.11.0

## Subdependencies
#certifi==2018.1.18
#chardet==3.0.4
#defusedxml==0.5.0
#idna==2.6
#oauthlib==2.0.7
#PyJWT==1.6.1
#python3-openid==3.1.0
#requests-oauthlib==0.8.0
#urllib3==1.22

## Bokeh Subdependencies:
#Jinja2==2.10
#MarkupSafe==1.0
#numpy==1.14.2
#packaging==17.1
#pyparsing==2.2.0
#PyYAML==3.12
#tornado==5.0.2

## Python Environment
#pip==18.0
#setuptools==39.0.1
#wheel==0.31.0
